requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
pypdf
google-generativeai
groq
python-multipart
psycopg2-binary
pgvector